um milhão de coisas: validação nas forms - correção de acentos - modificações de inteface - adição de funcionalidade

+++ src/main/java/br/com/fiap/starcrap/forms/PrincipalForm.java
@@ -9,12 +9,14 @@ import br.com.fiap.starcrap.daos.AlunoDAO;
 import br.com.fiap.starcrap.daos.EquipeDAO;
 import br.com.fiap.starcrap.models.Aluno;
 import br.com.fiap.starcrap.models.Equipe;
+import br.com.fiap.starcrap.models.Lancamento;
 
 import javax.persistence.EntityManager;
 import javax.persistence.Persistence;
 import javax.swing.*;
 import java.util.ArrayList;
 import java.util.List;
+import javax.validation.ConstraintViolationException;
 
 /**
  *
@@ -46,12 +48,12 @@ public class PrincipalForm extends javax.swing.JFrame {
     private void initComponents() {
 
         jPanel2 = new javax.swing.JPanel();
-        grupoButton = new javax.swing.JButton();
-        removerButton = new javax.swing.JButton();
         grupoScroll = new javax.swing.JScrollPane();
         grupoList = new javax.swing.JList();
         lancamentoButton = new javax.swing.JButton();
-        jButton1 = new javax.swing.JButton();
+        consultarButton = new javax.swing.JButton();
+        grupoButton = new javax.swing.JButton();
+        removerButton = new javax.swing.JButton();
         jPanel1 = new javax.swing.JPanel();
         componentesScroll = new javax.swing.JScrollPane();
         componentesList = new javax.swing.JList();
@@ -60,22 +62,9 @@ public class PrincipalForm extends javax.swing.JFrame {
         buttonPesquisar = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+        setTitle("StarCrap");
 
         jPanel2.setBorder(javax.swing.BorderFactory.createTitledBorder(null, "Grupo", javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Dialog", 1, 14))); // NOI18N
-
-        grupoButton.setText("Adicionar");
-        grupoButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                grupoButtonActionPerformed(evt);
-            }
-        });
-
-        removerButton.setText("Remover");
-        removerButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                removerButtonActionPerformed(evt);
-            }
-        });
 
         grupoList.setToolTipText("");
         grupoList.addListSelectionListener(new javax.swing.event.ListSelectionListener() {
@@ -93,8 +82,28 @@ public class PrincipalForm extends javax.swing.JFrame {
             }
         });
 
-        jButton1.setText("Consultar lançamento");
-        jButton1.setEnabled(false);
+        consultarButton.setText("Consultar lançamento");
+        consultarButton.setEnabled(false);
+        consultarButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                consultarButtonActionPerformed(evt);
+            }
+        });
+
+        grupoButton.setText("Adicionar");
+        grupoButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                grupoButtonActionPerformed(evt);
+            }
+        });
+
+        removerButton.setText("Remover");
+        removerButton.setEnabled(false);
+        removerButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                removerButtonActionPerformed(evt);
+            }
+        });
 
         javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
         jPanel2.setLayout(jPanel2Layout);
@@ -103,28 +112,28 @@ public class PrincipalForm extends javax.swing.JFrame {
             .addGroup(jPanel2Layout.createSequentialGroup()
                 .addContainerGap()
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                        .addGroup(jPanel2Layout.createSequentialGroup()
-                            .addComponent(grupoButton)
-                            .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                            .addComponent(removerButton))
-                        .addComponent(grupoScroll, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.PREFERRED_SIZE, 201, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(grupoScroll, javax.swing.GroupLayout.PREFERRED_SIZE, 201, javax.swing.GroupLayout.PREFERRED_SIZE)
                     .addComponent(lancamentoButton, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(jButton1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                    .addComponent(consultarButton, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
+                        .addGap(0, 0, Short.MAX_VALUE)
+                        .addComponent(grupoButton)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(removerButton)))
                 .addContainerGap())
         );
         jPanel2Layout.setVerticalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
-                .addComponent(grupoScroll, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(grupoButton)
                     .addComponent(removerButton))
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(grupoScroll, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(lancamentoButton)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jButton1)
+                .addComponent(consultarButton)
                 .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
@@ -186,26 +195,27 @@ public class PrincipalForm extends javax.swing.JFrame {
             .addGroup(layout.createSequentialGroup()
                 .addContainerGap()
                 .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(18, 18, 18)
                 .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
-                        .addGap(0, 14, Short.MAX_VALUE)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(18, 18, 18)
                         .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(12, 12, 12)))
-                .addContainerGap())
+                        .addContainerGap(22, Short.MAX_VALUE))
+                    .addGroup(layout.createSequentialGroup()
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                        .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                        .addContainerGap())))
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(layout.createSequentialGroup()
-                .addContainerGap()
-                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING)
                     .addGroup(layout.createSequentialGroup()
-                        .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 155, javax.swing.GroupLayout.PREFERRED_SIZE)
-                        .addGap(26, 26, 26)
+                        .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, 164, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                         .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
                     .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, 255, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
         );
 
         jPanel2.getAccessibleContext().setAccessibleName("Grupo");
@@ -214,55 +224,60 @@ public class PrincipalForm extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void grupoButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_grupoButtonActionPerformed
-        manager.getTransaction().begin();
+        try {
+            manager.getTransaction().begin();
 
-        String grupo = JOptionPane.showInputDialog("Digite o nome do grupo:");
-        if (grupo != null) {
-            Equipe equipe = new Equipe();
-            equipe.setNome(grupo);
+            String grupo = JOptionPane.showInputDialog(this, "Digite o nome do grupo:", "Cadastro de grupo",JOptionPane.PLAIN_MESSAGE);
+            if (grupo != null) {
+                Equipe equipe = new Equipe();
+                equipe.setNome(grupo);
 
-            String turma = JOptionPane.showInputDialog("Digite o nome da turma:");
+                String turma = JOptionPane.showInputDialog(this, "Digite o nome da turma:", "Cadastro de grupo",JOptionPane.PLAIN_MESSAGE);
 
-            if (turma != null) {
-                equipe.setTurma(turma);
-                int nAlunos = 0;
-                try {
-                    nAlunos = Integer.parseInt(JOptionPane.showInputDialog("Digite o número de alunos do grupo:"));
-                } catch (NumberFormatException e) {
-                    JOptionPane.showMessageDialog(null, "Digite somente números inteiros");
-                }
-                if (nAlunos > 0) {
-                    List<Aluno> alunos = new ArrayList<>();
+                if (turma != null) {
+                    equipe.setTurma(turma);
+                    int nAlunos = 0;
+                    try {
+                        nAlunos = Integer.parseInt(JOptionPane.showInputDialog(this,"Digite o número de alunos do grupo:", "Cadastro de grupo",JOptionPane.PLAIN_MESSAGE));
+                    } catch (NumberFormatException e) {
+                        JOptionPane.showMessageDialog(this, "Digite somente números inteiros", "Atenção!",JOptionPane.WARNING_MESSAGE);
+                    }
+                    if (nAlunos > 0) {
+                        List<Aluno> alunos = new ArrayList<>();
 
-                    for (int k = 0; k < nAlunos; k++) {
-                        Aluno aluno = new Aluno();
-                        String alunome = JOptionPane.showInputDialog("Digite o nome do " + (k + 1) + "º aluno:");
-                        if (alunome == null) {
-                            break;
-                        } else {
-
-                            String rm = JOptionPane.showInputDialog("Digite o rm do aluno:");
-                            if (rm == null) {
+                        for (int k = 0; k < nAlunos; k++) {
+                            Aluno aluno = new Aluno();
+                            String alunome = JOptionPane.showInputDialog(this, "Digite o nome do " + (k + 1) + "º aluno:", "Cadastro de grupo",JOptionPane.PLAIN_MESSAGE);
+                            if (alunome == null) {
                                 break;
                             } else {
-                                aluno.setNome(alunome);
-                                aluno.setRm(rm);
 
-                                alunoDAO.insert(aluno);
+                                String rm = JOptionPane.showInputDialog(this, "Digite o rm do aluno:", "Cadastro de grupo",JOptionPane.PLAIN_MESSAGE);
+                                if (rm == null) {
+                                    break;
+                                } else {
+                                    aluno.setNome(alunome);
+                                    aluno.setRm(rm);
 
-                                alunos.add(aluno);
+                                    alunoDAO.insert(aluno);
+
+                                    alunos.add(aluno);
+                                }
                             }
                         }
-                    }
 
-                    equipe.setAlunos(alunos);
-                    equipeDAO.insert(equipe);
-                    manager.getTransaction().commit();
-                    populaGrupoList();
+                        equipe.setAlunos(alunos);
+                        equipeDAO.insert(equipe);
+                        manager.getTransaction().commit();
+                        populaGrupoList();
+                    }
                 }
             }
+            fechaTransacao();
+        } catch (ConstraintViolationException e) {
+            JOptionPane.showMessageDialog(this, "Não foi possível inserir os dados no banco, verifique os valores inseridos (violação de constraint).", "Violação!",JOptionPane.ERROR_MESSAGE);
+            fechaTransacao();
         }
-        fechaTransacao();
     }//GEN-LAST:event_grupoButtonActionPerformed
 
     private void grupoListValueChanged(javax.swing.event.ListSelectionEvent evt) {//GEN-FIRST:event_grupoListValueChanged
@@ -272,6 +287,8 @@ public class PrincipalForm extends javax.swing.JFrame {
             numComponentes.setText("(" + String.valueOf(equipeSelecionada.getAlunos().size()) + ")");
             if (!lancamentoButton.isEnabled()) {
                 lancamentoButton.setEnabled(true);
+                consultarButton.setEnabled(true);
+                removerButton.setEnabled(true);
             }
         } else {
             numComponentes.setText("");
@@ -279,12 +296,18 @@ public class PrincipalForm extends javax.swing.JFrame {
     }//GEN-LAST:event_grupoListValueChanged
 
     private void removerButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_removerButtonActionPerformed
-        manager.getTransaction().begin();
-        equipeDAO.delete(equipeSelecionada);
-        manager.getTransaction().commit();
-        populaGrupoList();
-        limpaComponentesList();
-        fechaTransacao();
+        if(JOptionPane.showConfirmDialog(this, "Tem certeza de que deseja remover o grupo "+equipeSelecionada.getNome()+"?","Remoção de Grupo",JOptionPane.YES_NO_OPTION)==JOptionPane.YES_OPTION){
+        try {
+            manager.getTransaction().begin();
+            equipeDAO.delete(equipeSelecionada);
+            manager.getTransaction().commit();
+            populaGrupoList();
+            limpaComponentesList();
+            fechaTransacao();
+        } catch (ConstraintViolationException e) {
+            JOptionPane.showMessageDialog(this, "Não foi possível inserir os dados no banco de dados, verifique os valores inseridos (violação de constraint).","Violação!",JOptionPane.ERROR_MESSAGE);
+        }
+      }
     }//GEN-LAST:event_removerButtonActionPerformed
 
     private void lancamentoButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_lancamentoButtonActionPerformed
@@ -298,6 +321,37 @@ public class PrincipalForm extends javax.swing.JFrame {
         pesquisar.setVisible(true);
         this.dispose();
     }//GEN-LAST:event_buttonPesquisarActionPerformed
+
+    private void consultarButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_consultarButtonActionPerformed
+        if (equipeSelecionada.getLancamentos() == null || equipeSelecionada.getLancamentos().isEmpty()) {
+            JOptionPane.showMessageDialog(this, "Não há lançamentos para a equipe " + equipeSelecionada.getNome(),"Ops!",JOptionPane.INFORMATION_MESSAGE);
+            return;
+        }
+        Lancamento lanca = equipeSelecionada.getLancamentos().get(0);
+        int dia = lanca.getDataDoLancamento().getDayOfMonth();
+        int mes = lanca.getDataDoLancamento().getMonthValue();
+        int ano = lanca.getDataDoLancamento().getYear();
+
+        String str = "ID:" + lanca.getId() + "  -  Data: " + dia + "/" + mes + "/" + ano
+                + "\nVelocidade do Vento: " + lanca.getVelocidadeDoVento()
+                + "\nAltitude Máxima: " + lanca.getAltitudeMaxima()
+                + "\nVelocidade máxima: " + lanca.getVelocidadeMaxima()
+                + "\nTempo de Propulsão: " + lanca.getTempoDePropulsao().getSeconds()
+                + "\nPico de Aceleração: " + lanca.getPicoDeAceleracao()
+                + "\nAceleração Média: " + lanca.getAceleracaoMedia()
+                + "\nTempo entre Apogeu e Descida: " + lanca.getTempoEntreApogeuEDescida().getSeconds()
+                + "\nTempo de Ejeção: " + lanca.getTempoDeEjecao().getSeconds()
+                + "\nAltitude de Ejeção: " + lanca.getAltitudeDeEjecao()
+                + "\nTaxa de Descida: " + lanca.getTaxaDeDescida()
+                + "\nDuração do Vôo: " + lanca.getDuracaoDoVoo().getSeconds()
+                + "\n\nEspecificações do Foguete: "
+                + "\nID: " + lanca.getFoguete().getId()
+                + "\nDistância do Alvo: " + lanca.getFoguete().getDistanciaDoAlvo()
+                + "\nÂngulo de Lançamento: " + lanca.getFoguete().getAnguloDeLancamento()
+                + "\nPeso do Foguete: " + lanca.getFoguete().getPesoDoFoguete();
+
+        JOptionPane.showMessageDialog(this, str,"Dados de lançamento da equipe"+equipeSelecionada.getNome(),JOptionPane.INFORMATION_MESSAGE);
+    }//GEN-LAST:event_consultarButtonActionPerformed
 
     private void populaGrupoList() {
         DefaultListModel<Equipe> listModel = new DefaultListModel<>();
@@ -336,10 +390,10 @@ public class PrincipalForm extends javax.swing.JFrame {
     private javax.swing.JButton buttonPesquisar;
     private javax.swing.JList componentesList;
     private javax.swing.JScrollPane componentesScroll;
+    private javax.swing.JButton consultarButton;
     private javax.swing.JButton grupoButton;
     private javax.swing.JList grupoList;
     private javax.swing.JScrollPane grupoScroll;
-    private javax.swing.JButton jButton1;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JPanel jPanel2;
     private javax.swing.JPanel jPanel3;
